Version 8.3
Improved the city search function, now it gives more results and it accepts only the postal codes compatible with the API, i.e. those indicated in the help window by pressing F1.
Eliminated a check for the previous Yahoo API that pressing NVDA + shift + w gave an IndexError for a city with a local date greater or less than that of the system.
Added local date in city details.
Improved coding in Remove, Rename and Save cities (for NVDA with python 2)
Improved the forecast output windows, now the commands ctrl + a and ctrl + c also work and you can also close them with the enter key.